lib-pthread: Fix pthread_create
The last argument of pthread_create is opaque for the library/host.
It should be returned to the app as it is without any interpretations.
There is no point to translate it to a native address.
Actually it harms because the validation seems broken.

With this change, pthread_create/join seems working reasonably
with libc-wasi on macOS.