Avoid hardcoding the path of wasi-sdk in toolchain.cmake
It seems ${CMAKE_TOOLCHAIN_FILE} will be involved
multiple times and WASI_SDK_DIR will be treated as
undefined and in fact it was defined.

Somebody said the ${CMAKE_TOOLCHAIN_FILE} will not
always access cached variables. But it is not helpful
to store WASI_SDK_DIR in a local variable.