Fix pointer reuse when returning imports
Two of the branches when iterating imports in `wasm_module_imports` never set the `module_name` and `name` fields used when later calling `wasm_importtype_new`.

This leads to the same pointer being returned as part of multiple imports.
Freeing the returned vector then leads to double free, and freeing individual elements can lead to use after free.

Reusing the pointers is bad and causes undefined behavior, not to mention that the returned imports' names and module names are simply not correct, so we mimic the string copying done in the other branches to fix the issues mentioned above.