app-manager: Only excute main if wasi is enabled and app is wasi mode
If libc-wasi is not enabled, and the app is wanted to be loaded by both iwasm and host_tool, we can call on_init in main to make the app compatible when using these two tools. If libc-wasi is enabled, this modification will avoid main and on_init to be excuted both when using host_tool to install the app.