Add symbol check in wasm_c_api linkage
The orignal implementation does not consider the symbol (function name)
when linking an imported function. It can cause wrong linked function
when two function have same type. And it also introduce the unnecessary
limitation of the sequence that the external function declared in C must
be same with import sequence declared in wasm (a counter-example as
externs[] defined in callback_with_symbol.c when ).

This patch compare the symbol if it is specified in C. which could help
fix the aforementioned issues.