linux-sgx: allow to open files with arbitrary pathes in the sandbox using IPFS.
A limitation of the current implementation of SGX IPFS in WAMR prevented to open of
files which were not in the current directory. This restriction is lifted and can
now open files in paths, similarly to the WASI openat call, which takes into
account the sandbox of the file system.