Create module hash for each module in SGX lib-rats
Current SGX lib-rats wasm module hash is stored in a global buffer,
which may be overwritten if there are multiple wasm module loadings.
We move the module hash into the enclave module to resolve the issue.

And rename the RATS macro/variable in Makefile and Enclave.edl to
make the code more consistent.

And refine the sgx-ra sample document.