samples/wasi-threads: Fix wasm app target
As far as I know, cmake can't handle two different compilers and
targets within a project as this sample is trying to do.
This commit is a band-aid fix.

While CMAKE_C_COMPILER_TARGET here happens to work when both of
host compiler and wasm compiler are clang, it doesn't seem working
if the host compiler is gcc. (as it is on the linux jobs on the CI.)