Fix bh_assert for 64-bit platforms
In some cases (for me, every time I use memory sanitizer) the memory address
of some of the variables has 4 least significant bytes set to zero. Because
we cast the pointer to int, we look only at 4 (at least on my system, but I
guess on many other systems too) least significant bytes; the assertion fails
because 4 least significant bytes are 0.

I change bh_assert implementation so it casts to int64_t and it works well
with 64-bit architectures.