Add test for validating linear memory size updates
When memory.grow operation is called, it updates a `linear_mem_size` variable
in the `wasm_interp_call_func_bytecode()`; that update is not propagated to
all the threads, therefore if the function already is running and want to
access the memory that was added (after `memory.grow`) e.g. by calling
`memory.copy` where we check boundaries, the `memory.copy` fails with:
`Exception: out of bounds memory access` exception. I'm adding a test for now,
but will work on the implementation later.

In the meantime, suggestions on how to approach this problem are very welcome