set_exception_visitor: remove the special case for wasi proc exit
while wasi proc exit is not a real trap, what the runtime does on it
is mostly same as real traps. that is, kill the siblings threads
and represent the exit/trap as the result of the "process" to
the user api. there seems no reason to distinuguish it from
real traps here.

note that:

* the target thread either doesn't care the specific exception type
  or ignore wasi proc exit by themselves. (clear_wasi_proc_exit_exception)

* clear_wasi_proc_exit_exception only clears local exception.