Revert "Return error when exception was raised after main thread finishes (#2169)"
This reverts commit 71d43f3ca18b96c983d0fa8ca45ea658d6f9ee4a.

for some reasons, the original commit assumes a trap should win over
an exit even if the former is 1 second behind the latter. i don't
see any reason it should. in that case, the expected behavior is that
the exit likely wins and the other thread won't even reach the code to
raise a trap.