Add capability to filter WASI tests
In order to start running WASI tests in CI on Windows, add the ability
to filter WASI tests in the test runner scripts. The WASI libc
implementation is not yet complete on Windows so not all the tests are
passing. Once the implementation is complete, we can remove the filter
(although it's probably still useful to keep the filtering functionality
in the test runner either way).