Update Zephyr README.md to be generic towards all boards
Updates the Zephyr README.md to be generic towards all boards rather
than depending on a build script that can easily become stale. We may
opt to have a set of boards that we guarantee builds on in the future,
such as those for which board configuration files are currently
provided. At that time, we should ensure that those boards are tested in
CI.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>