Allow using mmap for shared linear memory if OS_ENABLE_HW_BOUND_CHECK is disabled
For shared memory, the max memory size must be defined in advanced. Re-allocation
for growing memory can't be used as it might change the base address, therefore for
OS_ENABLE_HW_BOUND_CHECK enabled the memory is mmaped, and if the flag is disabled,
the memory is allocated. This change introduces a flag that allows users to use mmap
for reserving memory address space even if the OS_ENABLE_HW_BOUND_CHECK is disabled