Fix arm64 issues on mac.
Make wamrc normalize "arm64" to "aarch64". Previously the only way to
make the "arm64" target was to not specify a target on 64 bit
arm-based mac builds. Now arm64 and aarch64 are treated as the same.

Make aot_loader accept "aarch64v8" on arm-based apple (as well as
accepting legacy "arm64" based aot targets).

This also removes __APPLE__ and __MACH__ from the block that defaults
size_level to 1 since