Track non-0x00 tableindex as ref types use
Any use of a table index that isn't exactly a null byte (`0x00`) means
that the module makes use of the reference types proposal. This is
important to track because `aot_compiler.c` will blindly assume that all
table indices are a single byte long otherwise.

This fixes a crash in WAMR for modules that contain multi-byte encodings
of table indices in `call_indirect` but make no other use of reference
types features.