CMakeLists.txt: Do not require C++
By default, the project() CMake command defaults to C and C++. [1]
Therefore, CMake might perform tests for both C and C++ compilers as
part of the configuration phase.

However, this has the consequence of the configuration phase to fail if
the system does not have a C++ toolchain installed, even if C++ is not
really used by the project.

[1]: https://cmake.org/cmake/help/latest/command/project.html