Add API for disabling native stack boundary checks
This API can be helpful if e.g. module is working in a reactor mode, and
the module is being used from multiple threads (where the caller guarantees
that no two threads execute code on the same execution environment).
Currently this usecase fails because the native stack boundaries for each
thread are different.
Alternative to this solution would be to update the boundary on every thread
switch but that might not be desired due to performance.