Make sure that touching pages will not fail due to misconfigured protection
When threads are initialized with `wasm_runtime_init_thread_env()`, the first
few pages (configured using `STACK_OVERFLOW_CHECK_GUARD_PAGE_COUNT`) are
configured to be not readable/writable (for the stack protection).
The protection is then removed in `destroy_stack_guard_pages()`, which is
automatically called when `os_thread_wrapper` is used for creating a new
thread.
However, the wrapper might not always be used (and in fact, is not in some
of the examples like [here](samples/terminate/src/main.c)) when the thread
is created outside of WAMR. To avoid crashes, we'll mark the whole stack
readable/writable before touching the pages.