Implement a thread-level-singleton WASMRuntime
Plan to replace process-level global variables with thread-local variables.
The reason for not using the variable attribute `__thread` is due to
concerns that it is a GCC extension and requires significant support from the
linker (ld), dynamic linker (ld.so), and system libraries
(libc.so and libpthread.so), making it not available everywhere.

Plus, memory initialization