docs: clarify that * in signatures does address conversion (#4681)
docs: clarify that * and $ in signatures does address conversion

It was non-obvious that `*` means the runtime does the app to native
conversion automatically. With the current docstring, one might assume
that the runtime only does boundary checks and not address conversion.
Under this assumption, passing the already-native address to
`wasm_runtime_addr_app_to_native` will lead to unexpected behavior